+++ src/main/java/com/backend/services/ClientService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.backend.entities.Client;
 import com.backend.repositories.ClientRepository;
+import com.backend.services.exceptions.DatabaseException;
+import com.backend.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class ClientService {
@@ -21,7 +24,7 @@ public class ClientService {
 	
 	public Client findById(Long id) {
 		Optional<Client> obj = repository.findById(id);
-		return obj.get();
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 	
 	public Client insert(Client obj) {
@@ -29,14 +32,26 @@ public class ClientService {
 	}
 	
 	public Client update(Client obj, Long id) {
-		Client client = findById(id);
-		client.setAccessionDate(obj.getAccessionDate());
-		client.setName(obj.getName());
-		client.setPhoneNumber(obj.getPhoneNumber());
-		return repository.save(client);
+		try {
+			if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+			Client client = findById(id);
+			client.setAccessionDate(obj.getAccessionDate());
+			client.setName(obj.getName());
+			client.setPhoneNumber(obj.getPhoneNumber());
+			return repository.save(client);
+		} catch (ResourceNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+			repository.deleteById(id);
+		} catch (ResourceNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}	
 	}
 }
