+++ src/main/java/com/backend/services/EnrollmentService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.backend.entities.Enrollment;
 import com.backend.repositories.EnrollmentRepository;
+import com.backend.services.exceptions.DatabaseException;
+import com.backend.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class EnrollmentService {
@@ -21,7 +24,7 @@ public class EnrollmentService {
 	
 	public Enrollment findById(Long id) {
 		Optional<Enrollment> enrollment = repository.findById(id);
-		return enrollment.get();
+		return enrollment.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 	
 	public Enrollment insert(Enrollment obj) {
@@ -29,16 +32,28 @@ public class EnrollmentService {
 	}
 	
 	public Enrollment update(Enrollment obj, Long id) {
-		Enrollment enrollment = findById(id);
-		enrollment.setAcademyPlan(obj.getAcademyPlan());
-		enrollment.setClient(obj.getClient());
-		enrollment.setEnd(obj.getEnd());
-		enrollment.setEnrollmentStatus(obj.getEnrollmentStatus());
-		enrollment.setStart(obj.getStart());
-		return repository.save(enrollment);	
+		try {
+			if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+			Enrollment enrollment = findById(id);
+			enrollment.setAcademyPlan(obj.getAcademyPlan());
+			enrollment.setClient(obj.getClient());
+			enrollment.setEnd(obj.getEnd());
+			enrollment.setEnrollmentStatus(obj.getEnrollmentStatus());
+			enrollment.setStart(obj.getStart());
+			return repository.save(enrollment);		
+		} catch (ResourceNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (!repository.existsById(id)) throw new ResourceNotFoundException(id);
+			repository.deleteById(id);			
+		} catch (ResourceNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 }
